requirements.txt
fastapi
uvicorn
pix2tex[gui]
kaleido
python-multipart
streamlit
Pillow